Use version-info instead of a build-receipt
The motivation behind this is that we want to be able to cache of the
Gradle build. The build receipt will
include a timestamp. In order to make most of the Jars cacheable
we would not include the build receipt in other jars but add it as a
different jar. For tests we use a different build receipt which doesn't
 include a timestamp or commitId.
 When building the intTestImage we replace the version info included
 in the distribution by the one used for tests, since the Gradle version
 for the test needs to be the same as for the distribution under test.

+review REVIEW-6331